Fix the mobile type

+++ models/voters/voters.go
@@ -58,7 +58,7 @@ type Voter struct {
 	Age                   int    `form:"age" json:"age,omitempty"`
 	Vote                  int    `form:"vote" json:"vote,omitempty"`
 	Email                 string `form:"email" json:"email,omitempty"`
-	Mobile_no             int    `form:"mobile_no" json:"mobile_no,omitempty"`
+	Mobile_no             int64  `form:"mobile_no" json:"mobile_no,omitempty"`
 	Image                 string `form:"image" json:"image,omitempty"`
 }
